4.05

+++ src/main/java/dal/AddlistDao.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface AddlistDao {
     void save(Addlist entity);
-    public void deleteTheAddList(int idCase);
+    public void deleteTheAddList(long idCase);
     public void update(Addlist entity);
      public List<Addlist> findByMainlistId(Long mainlistId);
 }
