+++ src/main/java/dal/MainlistDao.java
@@ -6,7 +6,7 @@ import java.util.List;
 public interface MainlistDao {
      void save(Mainlist entity);
     public List<Mainlist> findAll();
-    public void deleteTheMainList(int idCase);
+    public void deleteTheMainList(long  idCase);
     public void update(Mainlist entity);
     public List<Mainlist> findImportance();
     public Mainlist findid(long id);
